requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
msgpack==1.1.2
multidict==6.7.0
propcache==0.4.1
PyJWT==2.10.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
twilio==9.9.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
gunicorn
psycopg2-binary
dj-database-url
whitenoise 